Kazoo-Zookeeper/UI/requirements.txt
Flask
kazoo
requests
pyjwt
pytz